Safari-611.1.10/LayoutTests/fast/events/popup-blocking-timers6-expected.txt

Click Here
Test calling window.open() with a 1100 ms delay. A popup should not be allowed.
PASS newWindow is null
